Search export fix (task #4030)
Since last refactoring of the search functionality, we no longer
store the search results into the entity. Export functionality was
relying on those results to populate the csv data and export them
into a file. To fix this issue we run a new search based on the
saved search criteria to prepare the csv data to be exported.